Adding python for parsing jean.dat
The original Les Miserables data is more useful than I realized.  It is
a temporal graph in the sense that it contains detail data for every
chapter.  We can build and rebuild a graph based on a *range* of
chapters.  So I’m adding to dev a directory of stuff for working with
the data.